editor: integrate invenio-record-editor
Disables secondary bundling and uses directly the Webpack generated
bundles to support sourcemapped Sentry error logs.
