+++ inspirehep/modules/editor/templates/editor/index.html
@@ -1,0 +1,40 @@
+{#
+# This file is part of INSPIRE.
+# Copyright (C) 2017 CERN.
+#
+# Invenio is free software; you can redistribute it
+# and/or modify it under the terms of the GNU General Public License as
+# published by the Free Software Foundation; either version 2 of the
+# License, or (at your option) any later version.
+#
+# Invenio is distributed in the hope that it will be
+# useful, but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+# General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with Invenio; if not, write to the
+# Free Software Foundation, Inc., 59 Temple Place, Suite 330, Boston,
+# MA 02111-1307, USA.
+#
+# In applying this license, CERN does not
+# waive the privileges and immunities granted to it by virtue of its status
+# as an Intergovernmental Organization or submit itself to any jurisdiction.
+#}
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <title>Record Editor</title>
+    <link href="{{ url_for('static', filename='node_modules/record-editor/dist/styles.bundle.css') }}" rel="stylesheet"/>
+  </head>
+  <body>
+    <re-app>
+      Loading...
+    </re-app>
+
+    <script type="text/javascript" src="{{ url_for('static', filename='node_modules/record-editor/dist/inline.bundle.js') }}"></script>
+    <script type="text/javascript" src="{{ url_for('static', filename='node_modules/record-editor/dist/vendor.bundle.js') }}"></script>
+    <script type="text/javascript" src="{{ url_for('static', filename='node_modules/record-editor/dist/main.bundle.js') }}"></script>
+  </body>
+</html>
