+++ requirements.txt
@@ -23,8 +23,5 @@
 # Bleeding edge packages not yet released on Pypi
 -e git+https://github.com/CZ-NIC/python-rt.git@64cc54b1aa7514906be60f8bbd7e5ed18b56bba6#egg=rt
 
-# JSON editor
--e git+https://github.com/inveniosoftware-contrib/invenio-record-editor.git#egg=invenio-record-editor
-
 # SWORD2 client fork
 -e git+https://github.com/inspirehep/python-client-sword2.git#egg=sword2
